testsuite/gna: add a test for non-existing response file

+++ testsuite/gna/issue136/testsuite.sh
@@ -2,6 +2,16 @@
 
 . ../../testenv.sh
 
+if $GHDL @unknown.rsp 2> nexist.err; then
+  echo "failure expected"
+  exit 1
+fi
+if grep -q Bug nexist.err; then
+  echo "Bug box not expected"
+  exit 1
+fi
+rm -f nexist.err
+
 $GHDL @args.resp
 
 rm -rf html
